% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{add_simple_ld}
\alias{add_simple_ld}
\title{Create a new SNP in linkage disequiblibrium (LD) with a given SNP}
\usage{
add_simple_ld(causal_snp, r2)
}
\arguments{
\item{causal_snp}{A vector of genotypes for a single SNP}

\item{r2}{The r-squared value for LD between the new SNP and the causal SNP}
}
\value{
A vector of genotypes for the new SNP
}
\description{
Create a new SNP in linkage disequiblibrium (LD) with a given SNP
}
