% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{add_simple_ld_block}
\alias{add_simple_ld_block}
\title{Add a block of SNPs in LD with a given SNP}
\usage{
add_simple_ld_block(snp, block_size = 10)
}
\arguments{
\item{snp}{A vector of genotypes for a single SNP}

\item{block_size}{The number of SNPs in the block}
}
\value{
A matrix of genotypes for the block of SNPs
}
\description{
Add a block of SNPs in LD with a given SNP
}
