% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{binary_from_liability}
\alias{binary_from_liability}
\title{Convert a liability-scale phenotype to a binary phenotype}
\usage{
binary_from_liability(l, k = 0.2)
}
\arguments{
\item{l}{A liability-scale phenotype}

\item{k}{The prevalence. Default is 0.2.}
}
\value{
A binary phenotype
}
\description{
Requires a liability-scale phenotype and a prevalence (k). The function
will use the prevalence as a threshold to convert to a binary phenotype.
}
