% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{calculate_proportion}
\alias{calculate_proportion}
\title{Calculate proportion of regions covered by SNPs}
\usage{
calculate_proportion(n_in_regions, region_lengths)
}
\arguments{
\item{n_in_regions}{A vector of the number of SNPs in each region}

\item{region_lengths}{A vector of the lengths of each region}
}
\value{
A proportion of the regions covered by SNPs
}
\description{
Calculate proportion of regions covered by SNPs
}
\examples{
calculate_proportion(c(50, 40, 10), c(6000, 3000, 1000))

}
