% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.R
\name{call_liftover}
\alias{call_liftover}
\title{Liftover variants between builds}
\usage{
call_liftover(
  data,
  chr = "chr",
  pos = "pos",
  from_build = "grch37",
  to_build = "grch38",
  keep_original_positions = FALSE
)
}
\arguments{
\item{data}{A data frame with at least chromosome and position columns}

\item{chr}{The name of the column containing chromosome information. Default is "chr".}

\item{pos}{The name of the column containing position information. Default is "pos".}

\item{from_build}{The build to lift over from. Default is "grch37".}

\item{to_build}{The build to lift over to. Default is "grch38".}

\item{keep_original_positions}{A logical indicating whether to keep the original positions. Default is FALSE.}
}
\value{
A list containing the lifted data and a data frame of unmapped variants
}
\description{
This function takes a data frame of variants and lifts them over between builds.
Only supports liftover between grch37 and grch38. Unmapped variants are
discarded. By default, the original position is overwritten with the new
position. If keep_original_positions is set to TRUE, the original position
is retained in a new column as e.g. pos_grch37.
}
