% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{check_and_load_module}
\alias{check_and_load_module}
\title{Check and Load Module}
\usage{
check_and_load_module(module_name)
}
\arguments{
\item{module_name}{The name of the module to check and load.}
}
\description{
This function checks for the availability of a module on the system
and attempts to load it. It then checks if the module can be executed
successfully.
}
