% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{check_for_liftover}
\alias{check_for_liftover}
\title{Check for Liftover Files}
\usage{
check_for_liftover(attempt = 0)
}
\arguments{
\item{attempt}{The number of attempts made to download the liftover files. Default is 0.}
}
\description{
This function checks for the presence of the required liftover files for the package.
If the files are not present, it will attempt to download them by calling download_liftover().
If there has already been an attempt to download the files, it will stop and remove the tar.gz file.
}
