% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{check_if_snps_in_region}
\alias{check_if_snps_in_region}
\title{Check if any SNPs are in a given region}
\usage{
check_if_snps_in_region(
  snps_to_check,
  region_row,
  upstream_bp = 35000,
  downstream_bp = 10000
)
}
\arguments{
\item{snps_to_check}{A data frame containing the SNPs to check}

\item{region_row}{A row from a data frame containing the region to check}

\item{upstream_bp}{The number of base pairs upstream of the region to include}

\item{downstream_bp}{The number of base pairs downstream of the region to include}
}
\value{
The number of SNPs in the region
}
\description{
Check if any SNPs are in a given region
}
\examples{
region_row <- tibble::tibble(
 chr = 1,
 gene_start = 100000,
 gene_end = 200000
)
snps_to_check <- tibble::tibble(
 CHR = c(1, 1, 2, 3),
 POS = c(100000, 200000, 150000, 250000)
)
  
check_if_snps_in_region(snps_to_check, region_row)
  
 @export
}
