% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{check_tool_runs}
\alias{check_tool_runs}
\title{Check Tool Runs}
\usage{
check_tool_runs(tool_command, load_command = NULL)
}
\arguments{
\item{tool_command}{The command to check for on the system path.}
}
\description{
This function checks if a tool is available on the system path
and can be executed. It does this by running the tool with the
\code{--version} flag and checking the output.
}
