% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{download_liftover}
\alias{download_liftover}
\title{Install Required Liftover Files}
\usage{
download_liftover(attempt = 1)
}
\description{
This function downloads the required liftover files for the package,
including chain files and the max/linux executables.
}
