% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_supplement}
\alias{download_supplement}
\title{Download and Unzip Supplemental Files}
\usage{
download_supplement(path, temp_out_path = "")
}
\arguments{
\item{path}{The URL of the file to download.}

\item{temp_out_path}{Optional; the output path for the downloaded file.}
}
\value{
The path to the downloaded (and possibly unzipped) files.
}
\description{
Downloads a file from the specified path and unzips it if necessary. If the
output path is not specified, a temporary path is used.
}
