% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{downsample_individuals}
\alias{downsample_individuals}
\title{Randomly downsample individuals for a case-control study}
\usage{
downsample_individuals(
  X,
  y,
  n_cases = NULL,
  n_controls = NULL,
  is_transposed = FALSE
)
}
\arguments{
\item{X}{A matrix or big.matrix of individuals*genotypes}

\item{y}{A phenotype (binary)}

\item{n_cases}{The number of cases to downsample to. Default is NULL.}

\item{n_controls}{The number of controls to downsample to. Default is NULL.}

\item{is_transposed}{Whether the passed matrix is transposed (SNPs as rows,
individuals as columns) or not transposed (individuals as rows, SNPs as columns).}
}
\value{
A list with two elements:
\itemize{
\item [\link{1}] the downsampled matrix or big.matrix of individuals*genotypes (X)
\item [\link{2}] the downsampled binary phenotype (y)
}
}
\description{
Randomly downsample individuals for a case-control study
}
