% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exclude_apoe_region}
\alias{exclude_apoe_region}
\title{Exclude Genes in APOE Region}
\usage{
exclude_apoe_region(data)
}
\arguments{
\item{data}{A dataframe containing gene annotations.}
}
\value{
A dataframe excluding genes in the APOE region.
}
\description{
Filters out genes located in the APOE region based on predefined coordinates.
}
