% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{force_canonical_autosomes}
\alias{force_canonical_autosomes}
\title{Filter and Adjust Chromosome Data for Canonical Autosomes}
\usage{
force_canonical_autosomes(df)
}
\arguments{
\item{df}{A dataframe containing chromosome data as strings.}
}
\value{
A dataframe filtered for canonical autosomes with integer chromosome identifiers.
}
\description{
Filters a dataframe to include only canonical autosomes (1-22) and converts
chromosome identifiers to integers. Assumes input chromosomes are strings.
}
