% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{get_coverage_for_chromosome}
\alias{get_coverage_for_chromosome}
\title{Get the coverage percentage for a chromosome based on a set of regions}
\usage{
get_coverage_for_chromosome(data, chr_reference, chrom = 1)
}
\arguments{
\item{data}{A data frame containing the regions}

\item{chr_reference}{A data frame containing the reference genome coordinates}

\item{chrom}{The chromosome to calculate coverage for}
}
\value{
The coverage percentage for the chromosome
}
\description{
Get the coverage percentage for a chromosome based on a set of regions
}
\examples{
\dontrun{
  ref_coords <- read_ref_genome_coordinates()
  data <- sample_dummy_regions(n_regions = 100, min_length = 1000, max_length = 2000)
  get_coverage_for_chromosome(data, ref_coords, chrom = 1)
}

}
