% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_database_dir}
\alias{get_database_dir}
\title{Get Install Directory from Database TOML File}
\usage{
get_database_dir()
}
\description{
This function reads the databases.toml file and returns the installation path.
Will use the local data directory if the installation path is set to "data".
}
