% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_to_coordinates.R
\name{get_regions_from_genes}
\alias{get_regions_from_genes}
\title{Get chromosomal coordinates for a list of genes.}
\usage{
get_regions_from_genes(genes, gene_format = "hgnc_symbol", build = "grch38")
}
\arguments{
\item{genes}{A character vector of gene names.}

\item{gene_format}{A character string specifying the gene format. Default is "hgnc_symbol".}

\item{build}{A character string specifying the genome build. Default is "grch38".}
}
\value{
A tibble with columns "gene", "chromosome", "start", "end", "strand".
}
\description{
Get chromosomal coordinates for a list of genes.
}
\examples{
\dontrun{
genes <- c("PABPC4;HEYL", "PCSK9")
flat_genes <- unlist(stringr::str_split(genes, ";"))
coords <- get_regions_from_genes(flat_genes, gene_format = "hgnc_symbol")
}

}
