% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_zip}
\alias{is_zip}
\title{Check if the File is a Zip Archive}
\usage{
is_zip(f)
}
\arguments{
\item{f}{The file name to check.}
}
\value{
\code{TRUE} if the file is a zip archive, otherwise \code{FALSE}.
}
\description{
Determines whether the specified file has a \code{.zip} extension.
}
\examples{
is_zip("example.zip") # returns TRUE
is_zip("example.txt") # returns FALSE
}
