% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_files}
\alias{list_files}
\title{Generate a List of Files}
\usage{
list_files(range, prepend, append)
}
\arguments{
\item{range}{A vector of values to create file names for.}

\item{prepend}{Text to prepend to each file name.}

\item{append}{Text to append to each file name.}
}
\value{
A vector of constructed file names.
}
\description{
Creates a list of file names based on a specified range and naming pattern.
}
