% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_organisms.R
\name{mouse_to_human}
\alias{mouse_to_human}
\title{Convert Mouse Genes to Human Genes}
\usage{
mouse_to_human(genes)
}
\arguments{
\item{genes}{A character vector of mouse gene symbols (MGI symbols) to be mapped to human.}
}
\value{
A tibble containing the mapped human genes and their attributes.
The columns include "mgi_symbol" for the mouse gene symbol and
additional columns specified by \code{ensemble_columns_to_extract} and
\code{clean_output_name}. Only genes located on canonical autosomes are included.
}
\description{
This function maps mouse genes to their human orthologs using BioMart data. It retrieves
specified attributes for the mapped human genes, including chromosome locations, and
filters the results to include only canonical autosomes.
}
\examples{
\dontrun{
mouse_genes <- c("apoc1", "trem2") # Example mouse gene symbols
human_orthologs <- mouse_to_human(mouse_genes)
print(human_orthologs)
}

}
