% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_host}
\alias{parse_host}
\title{Parse a host path for a mart object}
\usage{
parse_host(genome, host = NULL)
}
\arguments{
\item{genome}{The required genome. Should be GRch37, GRCh38 or GRCm39.}

\item{host}{The suggested path. If NULL, lookup address, else check passed address.}
}
\value{
A string for the required host path
}
\description{
Looks up or parses the host address for a needed genome build.
}
