% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paste_names}
\alias{paste_names}
\title{Concatenate Names with Prefix and Suffix}
\usage{
paste_names(i, pre, post)
}
\arguments{
\item{i}{Index to be included between prefix and suffix.}

\item{pre}{Prefix for the name.}

\item{post}{Suffix for the name.}
}
\value{
A string resulting from the concatenation of prefix, index, and suffix.
}
\description{
Concatenates a given index with prefix and suffix strings.
}
