% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid_to_gene.R
\name{query_opentargets}
\alias{query_opentargets}
\title{Submit a query string to the Open Targets Genetics GraphQL API}
\usage{
query_opentargets(query_string, variables = NULL, wait = 0)
}
\arguments{
\item{query_string}{A string containing the query to be submitted to the Open Targets Genetics GraphQL API}

\item{variables}{A list of variables to be passed to the query}

\item{wait}{A numeric value indicating the number of seconds to wait before submitting the query}
}
\value{
A list containing the results of the query
}
\description{
Submit a query string to the Open Targets Genetics GraphQL API
}
