% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_bim_file}
\alias{read_bim_file}
\title{Read BIM File}
\usage{
read_bim_file(f)
}
\arguments{
\item{f}{The filepath to the .bim file.}
}
\value{
A dataframe with SNP information from the .bim file.
}
\description{
Reads a .bim file and extracts SNP information.
}
