% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_config}
\alias{read_config}
\title{Read Configuration File}
\usage{
read_config()
}
\value{
A list containing the parsed TOML configuration file
}
\description{
Reads a TOML configuration file containing package-wide constants
}
