% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_database_toml}
\alias{read_database_toml}
\title{Read Database TOML file}
\usage{
read_database_toml()
}
\value{
A list containing the parsed TOML configuration file
}
\description{
Reads a TOML configuration file containing download links and version numbers for external databases
}
