% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_ref_genome_coordinates}
\alias{read_ref_genome_coordinates}
\title{Read Reference Genome Coordinates}
\usage{
read_ref_genome_coordinates()
}
\value{
A dataframe containing reference genome coordinates.
}
\description{
Reads hardcoded reference genome coordinates, specifically for GRCh38.p14.
}
