% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_regions_to_search}
\alias{read_regions_to_search}
\title{Read Regions to Search for Enrichment}
\usage{
read_regions_to_search(f)
}
\arguments{
\item{f}{The filepath to the regions file or "dummy_region" for a test region.}
}
\value{
A dataframe of regions with calculated gene length in base pairs.
}
\description{
Reads in a file specifying regions to search for enrichment analysis, or
generates a dummy region for testing. Adds gene length as a column.
}
