% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.R
\name{read_unmapped_variants}
\alias{read_unmapped_variants}
\title{Read in unmapped variants}
\usage{
read_unmapped_variants(filename)
}
\arguments{
\item{filename}{The name of the file containing unmapped variants}
}
\value{
A data frame with chromosome, position0, and position1 columns
}
\description{
This function reads in a file of unmapped variants and returns a data frame,
where the unmapped variants are output generated by the liftover tool.
}
