% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{replace_with_ld}
\alias{replace_with_ld}
\title{Replace a proportion of causal SNPs with new SNPs in LD}
\usage{
replace_with_ld(g, m, ld_range = c(0.1, 0.9))
}
\arguments{
\item{g}{A matrix of genotypes}

\item{m}{The proportion of causal SNPs to keep}

\item{ld_range}{The range of r-squared values for LD between the new SNPs and the causal SNPs}
}
\value{
A matrix of genotypes with a proportion of causal SNPs replaced by new SNPs in LD
}
\description{
Replace a proportion of causal SNPs with new SNPs in LD
}
