% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid_to_gene.R
\name{rsids_to_functional_genes}
\alias{rsids_to_functional_genes}
\title{Convert a list of rsids to a list of genes and associated functional annotations from Open Targets Genetics}
\usage{
rsids_to_functional_genes(
  rsids,
  spacing_min = 0,
  spacing_max = 1,
  batch_size = 50,
  batch_spacing = 5
)
}
\arguments{
\item{rsids}{A character vector of rsids}

\item{spacing_min}{A numeric value indicating the minimum number of seconds to wait between requests}

\item{spacing_max}{A numeric value indicating the maximum number of seconds to wait between requests}

\item{batch_size}{A numeric value indicating the number of rsids to be submitted in each batch}

\item{batch_spacing}{A numeric value indicating the number of seconds to wait between batches}
}
\value{
A list containing (index 1), the annotated genes and (index 2), the rsids that were not found in the Open Targets Genetics database
}
\description{
This function is a wrapper to call both rsid_to_varid and varid_to_genes.
See \code{?varid_to_genes} for more details on the output.
}
