% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{sample_dummy_regions}
\alias{sample_dummy_regions}
\title{Sample multiple regions from the reference genome}
\usage{
sample_dummy_regions(n_regions = 100, min_length = 1000, max_length = 5e+05)
}
\arguments{
\item{n_regions}{The number of regions to sample}

\item{min_length}{The minimum length of a region}

\item{max_length}{The maximum length of a region}
}
\value{
A data frame containing the sampled regions
}
\description{
Sample multiple regions from the reference genome
}
\examples{
\dontrun{
  sample_dummy_regions(n_regions = 10, min_length = 1000, max_length = 2000)
}

}
