% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{sample_region}
\alias{sample_region}
\title{Sample a region from the reference genome}
\usage{
sample_region(region_length, chr_reference)
}
\arguments{
\item{region_length}{The length of the region to sample}

\item{chr_reference}{A data frame containing the reference genome coordinates}
}
\value{
A vector containing the chromosome, start, and end of the region
}
\description{
Sample a region from the reference genome
}
\examples{
ref_coords <- read_ref_genome_coordinates()
sample_region(1000, ref_coords)

}
