% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{sample_regions_from_list}
\alias{sample_regions_from_list}
\title{Sample multiple regions from the reference genome given a vector of region lengths}
\usage{
sample_regions_from_list(region_lengths)
}
\arguments{
\item{region_lengths}{A vector of the lengths of the regions to sample}
}
\value{
A data frame containing the sampled regions
}
\description{
Sample multiple regions from the reference genome given a vector of region lengths
}
\examples{
sample_regions_from_list(c(1000, 2000, 3000))

}
