% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{set_database_dir}
\alias{set_database_dir}
\title{Set Database Directory}
\usage{
set_database_dir(path, recursive = FALSE)
}
\arguments{
\item{path}{The full path to the directory to store files in.}

\item{recursive}{Whether to create any parent directories if they do not exist.}
}
\description{
Creates a directory for holding files and updates the path in the databases.toml file.
Useful where the install for R code is in a home directory but large database
files need to be stored elsewhere. Note that the location should be accessible
from compute nodes in the cluster as it will be accessed during jobs.
}
