% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_human_genome}
\alias{set_human_genome}
\title{Set Up Human Genome Annotation}
\usage{
set_human_genome(human_build = "grch38", host = NULL)
}
\arguments{
\item{human_build}{The human genome build to use, either "grch38" or "grch37".}

\item{host}{The ensembl address for the genome. Default is NULL.}
}
\value{
A BioMart object configured for the specified human genome build.
}
\description{
Configures the human genome annotation using BioMart based on the specified genome build.
}
