% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_mart}
\alias{set_mart}
\title{Set up BioMart connection}
\usage{
set_mart(dataset, host)
}
\arguments{
\item{dataset}{The dataset to connect to within the BioMart database.}

\item{host}{The URL of the BioMart host to connect to.}
}
\value{
An object of class Mart representing the BioMart connection.
}
\description{
Initializes a connection to a specified BioMart database and dataset.
}
\examples{
\dontrun{
  mart <- set_mart(dataset = "hsapiens_gene_ensembl", host = "http://www.ensembl.org/biomart")
}

}
