% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_bioinformatics_tools}
\alias{setup_bioinformatics_tools}
\title{Setup Bioinformatics Tools}
\usage{
setup_bioinformatics_tools(tools, hpc = FALSE, container_paths = list())
}
\arguments{
\item{tools}{A character vector of tool names to check for.}

\item{hpc}{If TRUE, checks modules can be loaded with appropriate tools first.}

\item{container_paths}{A named list of Singularity container paths for tools.}
}
\description{
This function checks for the availability of required bioinformatics tools
on the system path. If the tools are not available, it will attempt to load
the tool as a module if the \code{hpc} argument is set to TRUE. If the module
loading fails and a Singularity container is provided, it will attempt to
use the container.
}
