% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_complex_ld_blocks}
\alias{simulate_complex_ld_blocks}
\title{Simulate a matrix of genotypes with a proportion of causal SNPs in LD}
\usage{
simulate_complex_ld_blocks(
  p = 10,
  n = 1000,
  n_blocks = 10,
  block_size_range = c(3, 10)
)
}
\arguments{
\item{p}{The number of SNPs}

\item{n}{The number of individuals}

\item{n_blocks}{The number of LD blocks. Default is 10.}

\item{block_size_range}{The range of block sizes. Default is c(3, 10).}
}
\value{
A matrix of genotypes with a proportion of causal SNPs in LD
}
\description{
Will create n_blocks LD blocks of a size sample uniformly from the
range block_size_range.
}
