% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_e2_e4}
\alias{simulate_e2_e4}
\title{Simulate APOE alleles and APOE score}
\usage{
simulate_e2_e4(n)
}
\arguments{
\item{n}{The number of individuals to simulate}
}
\value{
A tibble with columns for e2, e4, and e2_e4_score
}
\description{
This function will simulate APOE alleles and APOE score based on the
allele frequencies and effect sizes from the config.toml file.
The calculated APOE score is the sum of the product of the allele
counts and effect sizes for e2 and e4.
}
