% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_y}
\alias{simulate_y}
\title{Simulate a phenotype give a matrix of genotypes}
\usage{
simulate_y(
  X,
  p_causal = 0.01,
  h2_l = 0.2,
  k = 0.2,
  n_cases = NULL,
  n_controls = NULL,
  verbose = TRUE
)
}
\arguments{
\item{X}{A matrix of genotypes}

\item{p_causal}{The proportion of causal SNPs. Default is 0.01.}

\item{h2_l}{The heritability on the liability scale. Default is 0.2.}

\item{k}{The prevalence. Default is 0.2.}

\item{n_cases}{The number of cases. Default is NULL.}

\item{n_controls}{The number of controls. Default is NULL.}

\item{verbose}{Whether to print the expected and actual heritability. Default is TRUE.}
}
\value{
A list with three elements: the matrix of genotypes, the phenotype and a logical vector indicating
which SNPs are causal.
}
\description{
For a given heritability (on the liability scale) and a proportion of causal
SNPs, this function will simulate a phenotype based on the genotypes. Note
that prevalence, k, and proportion causal SNPs, m, will both influence
how easy it is to predict the phenotype from the genotypes, not just
heritability. If n_cases and n_controls are not NULL, the function will
downsample the cases and controls to the desired numbers.
}
