% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.R
\name{to_bed}
\alias{to_bed}
\title{Create a BED file from a data frame}
\usage{
to_bed(data, chrom = "chr", position = "pos")
}
\arguments{
\item{data}{A data frame with at least chromosome and position columns}

\item{chrom}{The name of the column containing chromosome information. Default is "chr".}

\item{position}{The name of the column containing position information. Default is "pos".}
}
\value{
A data frame with chromosome, position0, and position columns
}
\description{
This function takes a data frame and creates a BED file from it.
Requires only chromosome and position to do so.
Note this assumes that for all positions you are working with
SNPs and not indels, so here all positions are 1bp long.
}
