% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{try_mart}
\alias{try_mart}
\title{Attempt to establish BioMart connection with error handling}
\usage{
try_mart(dataset, host)
}
\arguments{
\item{dataset}{The dataset to connect to within the BioMart database.}

\item{host}{The URL of the BioMart host to connect to initially.}
}
\value{
An object of class Mart representing the BioMart connection, or
\code{NA} if the connection could not be established after error handling.
}
\description{
Tries to connect to a specified BioMart database and dataset, with
error handling for HTTP connection issues. If the initial connection fails
due to a server-side error, it attempts to connect to a mirror site.
}
\examples{
\dontrun{
  mart <- try_mart(dataset = "hsapiens_gene_ensembl", host = "http://www.ensembl.org/biomart")
}

}
