% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{update_bash_profile}
\alias{update_bash_profile}
\title{Update .bash_profile}
\usage{
update_bash_profile()
}
\description{
This function checks if the CLI install path is in the .bash_profile file.
If it is not, it will add the path to the file.
}
