% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{upsample_individuals_independently}
\alias{upsample_individuals_independently}
\title{Upsample individuals in a matrix of genotypes}
\usage{
upsample_individuals_independently(
  data,
  nrows,
  ncols,
  n_ticks = 100,
  is_transposed = FALSE
)
}
\arguments{
\item{data}{A matrix, data.frame or big.matrix of individuals*genotypes}

\item{nrows}{The number of rows to be in the output data}

\item{ncols}{The number of columns to be in the output data}

\item{n_ticks}{The number of ticks for the progress bar. Default is 100.}

\item{is_transposed}{Whether the data is transposed (SNPs as rows,
individuals as columns), or not transposed (individuals as rows,
SNPs are columns). Default is FALSE.}
}
\value{
A big.matrix of upsampled genotypes of size nrows*ncols.
}
\description{
This function will upsample individuals in a matrix of genotypes. It will
sample with replacement from the existing individuals to create new
individuals. All SNPs are sampled independent instead of as a group,
which results in decorrelated SNPs. Note that is_transposed must be
set to TRUE if a matrix/data.frame is passed in the classic PLINK
format (SNPs as rows, individuals as columns). Note also that nrows
and ncols passed have to be swapped if is_transposed is changed too.
}
