% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_singularity_container}
\alias{use_singularity_container}
\title{Use Singularity Container}
\usage{
use_singularity_container(container_path)
}
\arguments{
\item{container_path}{The path to the Singularity container file.}
}
\description{
This function provides guidance on using a Singularity container for a required tool.
Function not currently supported
}
