% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid_to_gene.R
\name{varid_to_genes}
\alias{varid_to_genes}
\title{Convert a variant ID to a list of genes and associated functional annotations from Open Targets Genetics}
\usage{
varid_to_genes(variantId, wait = 0)
}
\arguments{
\item{variantId}{A character vector of variant IDs (formatted as chrom_pos_ref_alt)}

\item{wait}{A numeric value indicating the number of seconds to wait before submitting the query}
}
\value{
A tibble containing the results of the query
}
\description{
The function will return 0 or more rows for each variant.
0 for variants that are not found in the Open Targets Genetics database,
1 for single matches (where it has returned the best prediction from Open Targets)
And more than one if there are multiple genes within 1 s.d. of the top gene,
or if there are multiple functional annotations or both functional and positional
annotations (e.g. QTl data annotates to one genes but the variant is intronic to another gene).
}
