% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{install_dependencies}
\alias{install_dependencies}
\title{Install Required Dependencies}
\usage{
install_dependencies(hpc = FALSE)
}
\arguments{
\item{hpc}{If TRUE, checks modules can be loaded with appropriate tools first.}
}
\description{
This function checks for and installs required Bioconductor packages for the package,
and any command-line tools needed, e.g. plink, bcftools etc.
Currently, it ensures that \code{biomaRt} is installed, using \code{BiocManager}.
}
\details{
The function first checks if \code{BiocManager} is installed and installs it if necessary.
Then, it checks for the presence of specified Bioconductor packages (\code{biomaRt}) and installs any that are missing.
This is particularly useful for setting up the package environment or ensuring that dependencies are met for new users.
}
\examples{
\dontrun{
install_dependencies()
}

}
