% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_genomes}
\alias{set_genomes}
\title{Set Up Human and Mouse Genome Annotations}
\usage{
set_genomes(host = NULL)
}
\arguments{
\item{host}{The ensembl address for the genome. Default is NULL.}
}
\value{
A list containing BioMart objects for human and mouse genomes.
}
\description{
Configures BioMart objects for both human and mouse genome annotations.
}
